Some more fixes

+++ monix-execution/shared/src/test/scala/monix/execution/AsyncSemaphoreSuite.scala
@@ -181,15 +181,15 @@ object AsyncSemaphoreSuite extends TestSuite[TestScheduler] {
     // Executing Futures on the global scheduler!
     import scala.concurrent.ExecutionContext.Implicits.global
 
-    def acquireN(semaphore: AsyncSemaphore, n: Int): Future[Unit] =
+    def acquireN(semaphore: AsyncSemaphore, n: Long): Future[Unit] =
       Future(semaphore.tryAcquireN(n)).flatMap {
         case true => Future.successful(())
         case false => acquireN(semaphore, n)
       }
 
-    def withPermitN[A](semaphore: AsyncSemaphore, n: Int)(f: () => Future[A]): Future[A] =
+    def withPermitN[A](semaphore: AsyncSemaphore, n: Long)(f: () => Future[A]): Future[A] =
       acquireN(semaphore, n).flatMap { _ =>
-        FutureUtils.transform[A, A](f(), r => { semaphore.releaseN(n.toLong); r })
+        FutureUtils.transform[A, A](f(), r => { semaphore.releaseN(n); r })
       }
 
     repeatTest(10) { () =>
@@ -202,7 +202,7 @@ object AsyncSemaphoreSuite extends TestSuite[TestScheduler] {
         allReleased.completeWith(semaphore.awaitAvailable(available))
 
         val futures = for (i <- 0 until count) yield {
-          withPermitN(semaphore, Math.floorMod(Random.nextInt(), 3) + 1) { () =>
+          withPermitN(semaphore, (Math.floorMod(Random.nextInt(), 3) + 1).toLong) { () =>
             Future(1).map { x =>
               assert(!allReleased.isCompleted, s"!allReleased.isCompleted (index $i)")
               x
@@ -218,7 +218,7 @@ object AsyncSemaphoreSuite extends TestSuite[TestScheduler] {
         assertEquals(r, count)
         assertEquals(semaphore.available(), available)
       }
-    }
+    }    
   }
 
   def repeatTest(n: Int)(f: () => Future[Unit])(implicit ec: ExecutionContext): Future[Unit] =
